requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
SQLAlchemy==1.4.46
markdown2==2.4.10
Werkzeug==2.2.3
python-dotenv==1.0.0
pymongo==4.3.3
Flask-WTF==1.1.1